extractors: smda: extract format and OS characteristics at all scopes
